% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_read_pos.R
\name{gl_read_pos}
\alias{gl_read_pos}
\title{Read .pos file to data frame}
\usage{
gl_read_pos(pos)
}
\arguments{
\item{pos}{the location of a *.pos data-logger file}
}
\value{

}
\description{
Converts .pos file to data frame for easy manipulation
}
