% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glt_geoid.R
\name{glt_geoid}
\alias{glt_geoid}
\title{Get geoid correction}
\usage{
glt_geoid(longitude, latitude)
}
\arguments{
\item{longitude}{longitude, single value or vector}

\item{latitude}{latitude, single value or vector}
}
\value{
geoid height, for the correction of values measured relative
 to a reference ellipsoid (GPS)
}
\description{
Get a geoid correction value
for a latitude longitude pair
}
\examples{
\dontrun{
correction <- glt_geoid(4, 50)
}
}
