% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glt_migtech_pamlr.R
\name{glt_migtech_pamlr}
\alias{glt_migtech_pamlr}
\title{Converts Migrate Tech data to PAMLr}
\usage{
glt_migtech_pamlr(path)
}
\arguments{
\item{path}{location of lux (and optional .deg)
Migrate Technologies data files}
}
\value{
A nested list of variables, ID (taken from
the .lux files) is returned as a top level list
item, within it contained the data as formated
according to
}
\description{
Data conversion to wrangle data into the
PAMLr data format for easy visualization
of geolocator data
}
