% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glt_read_pos.R
\name{glt_read_pos}
\alias{glt_read_pos}
\title{Read .pos file to data frame}
\usage{
glt_read_pos(pos, geoid = TRUE)
}
\arguments{
\item{pos}{the location of a *.pos data-logger file}

\item{geoid}{correct altitudes for geiod shape (default = TRUE)}
}
\value{

}
\description{
Converts .pos file to data frame for easy manipulation
}
