% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{hPa_to_meter}
\alias{hPa_to_meter}
\title{Converts altitudes in hPa in m}
\usage{
hPa_to_meter(hpa)
}
\arguments{
\item{hpa}{altitude expressed as hPa}
}
\value{
vector in meters
}
\description{
Uses the standard atmosphere as a reference.
Note that this method differs from the ERA5 reference
}
