% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glt_flatten_PAML.R
\name{glt_flatten_PAML}
\alias{glt_flatten_PAML}
\title{Flatten PAMLr data}
\usage{
glt_flatten_PAML(pamlr_data)
}
\arguments{
\item{pamlr_data}{pam data as a nested list}
}
\value{
returns a tidy tibble with activity, pressure, pitch and
 temperature data
}
\description{
Convert nested PAMLr (list) data to a tidy data frame.
}
